extract opensearch functionality into own files, better async task handling
