+++ db-connector.go
@@ -78,180 +78,6 @@ type ShuffleStorage struct {
 
 var maxCacheKeyLength = 250
 
-// Create ElasticSearch/OpenSearch index prefix
-// It is used where a single cluster of ElasticSearch/OpenSearch utilized by several
-// Shuffle instance
-// E.g. Instance1_Workflowapp
-func GetESIndexPrefix(index string) string {
-	prefix := os.Getenv("SHUFFLE_OPENSEARCH_INDEX_PREFIX")
-	if len(prefix) > 0 {
-		return fmt.Sprintf("%s_%s", prefix, index)
-	}
-
-	return index
-}
-
-func GetOpensearchBaseIndexes() []string {
-	return []string{
-		"workflowexecution",
-		"workflowexecution_live",
-		"datastore_ngram",
-		"org_cache",
-		"org_cache_revisions",
-		"notifications",
-		"shuffle_logs",
-		"environments",
-		"org_statistics",
-		"workflowapp",
-		"workflow",
-		"workflow_revisions",
-		"datastore_category",
-	}
-}
-
-// GetOpensearchRolloverIndexes returns the subset of base indexes that are
-// genuinely append-only (or, for workflowexecution, append-only AFTER the
-// hot/cold lifecycle split below) and therefore want alias + automatic
-// rollover + ISM retention.
-//
-// notifications and workflowexecution_live are stateful keyed stores where
-// the same _id is re-written in place (notifications toggle read/ignored
-// flags forever; workflowexecution_live holds in-flight executions before
-// they reach a terminal status) - OpenSearch always routes alias writes to
-// the current write-index generation, so rolling them would split an _id
-// across generations and silently produce duplicate documents. They stay on
-// a single backing index like the other keyed stores.
-func GetOpensearchRolloverIndexes() []string {
-	return []string{
-		"shuffle_logs",
-		// Content-addressed / fresh-id-per-write revision stores: a given _id is
-		// never rewritten once created, so alias+rollover is safe here (unlike
-		// workflowexecution_live/notifications, which mutate the same _id in place).
-		"workflow_revisions",  // _id = md5(name+id+actions+triggers+variables)
-		"org_cache_revisions", // _id = <org>_<key>[_<category>]_<uuid>
-		// workflowexecution is the ARCHIVE for confirmed-terminal
-		// executions (see execution_lifecycle.go). Nothing writes to it
-		// except archiveExecutionDocument, which resolves any existing copy of an
-		// execution_id to its concrete backing index before writing (rather than
-		// writing through the alias blindly), so rollover is safe here even though
-		// the same execution_id can in rare cases be revisited (see the unarchive
-		// path in writeExecutionDocument).
-		"workflowexecution",
-	}
-}
-
-// opensearchCoreMappings holds pragmatic field mappings applied only when an
-// index is created fresh (OpenSearch mappings are immutable after creation, so
-// existing deployments keep whatever mappings they already have). Keys are the
-// base index names from GetOpensearchBaseIndexes.
-//
-// id/ref fields are keyword, date/epoch fields are date (epoch_second), and
-// numeric counters/priorities are long so numeric sorting and filtering work.
-// The default strings_as_keywords dynamic template only handles strings; these
-// explicit types keep the mapping stable regardless of how a value is written.
-//
-// Large/binary blobs (images) are mapped with index:false + doc_values:false:
-// they stay retrievable via _source but are never added to the inverted index
-// or doc_values, avoiding index bloat and too-big-field failures. Add a field
-// only when its type is known to be stable and it is actually sorted/filtered
-// on (or when it must be excluded from indexing).
-//
-// created/edited/started_at on workflowexecution are deliberately "long", not
-// "date": these three fields are sorted across workflowexecution_live+archive
-// (started_at, in GetUnfinishedExecutions/GetAllWorkflowExecutions[V2]/
-// GetWorkflowRunsBySearch) or across archive generations (created/edited, in
-// findExecutionInArchive). OpenSearch stores "date" doc values internally as
-// epoch milliseconds regardless of the "format" annotation - format only
-// affects range-query parsing, not the raw value a sort clause returns. Any
-// pre-existing customer index created before this mapping existed has these
-// fields dynamically inferred as "long" (the app always wrote raw epoch
-// SECONDS as JSON numbers). Mapping them "date" here would make freshly
-// created generations sort in milliseconds while old/legacy generations sort
-// in seconds - since ms values are ~1000x larger, every doc in a
-// date-mapped generation would silently outrank every doc in a long-mapped
-// generation regardless of true chronological order.
-// Keeping these three fields "long" matches what already-deployed clusters
-// have and keeps sort units identical across every generation, old and new.
-// completed_at has no cross-generation/cross-index sort today (only a
-// same-index numeric range filter in the archival sweep), so it is left as
-// "date" for existing single-index range-query compatibility.
-var opensearchCoreMappings = map[string]map[string]interface{}{
-	"workflowexecution": {
-		"properties": map[string]interface{}{
-			"execution_id":  map[string]interface{}{"type": "keyword"},
-			"workflow_id":   map[string]interface{}{"type": "keyword"},
-			"execution_org": map[string]interface{}{"type": "keyword"},
-			"status":        map[string]interface{}{"type": "keyword"},
-			"created":       map[string]interface{}{"type": "long"},
-			"edited":        map[string]interface{}{"type": "long"},
-			"started_at":    map[string]interface{}{"type": "long"},
-			"completed_at":  map[string]interface{}{"type": "date", "format": "epoch_second"},
-			"priority":      map[string]interface{}{"type": "long"},
-		},
-	},
-	"notifications": {
-		"properties": map[string]interface{}{
-			"id":                  map[string]interface{}{"type": "keyword"},
-			"org_id":              map[string]interface{}{"type": "keyword"},
-			"user_id":             map[string]interface{}{"type": "keyword"},
-			"execution_id":        map[string]interface{}{"type": "keyword"},
-			"workflow_id":         map[string]interface{}{"type": "keyword"},
-			"org_notification_id": map[string]interface{}{"type": "keyword"},
-			"created_at":          map[string]interface{}{"type": "date", "format": "epoch_second"},
-			"updated_at":          map[string]interface{}{"type": "date", "format": "epoch_second"},
-			"amount":              map[string]interface{}{"type": "long"},
-			"image":               map[string]interface{}{"type": "keyword", "index": false, "doc_values": false},
-			"read":                map[string]interface{}{"type": "boolean"},
-			"ignored":             map[string]interface{}{"type": "boolean"},
-			"dismissable":         map[string]interface{}{"type": "boolean"},
-			"personal":            map[string]interface{}{"type": "boolean"},
-		},
-	},
-	"org_statistics": {
-		"properties": map[string]interface{}{
-			"org_id":                      map[string]interface{}{"type": "keyword"},
-			"last_cleared":                map[string]interface{}{"type": "date", "format": "epoch_second"},
-			"total_app_executions":        map[string]interface{}{"type": "long"},
-			"total_workflow_executions":   map[string]interface{}{"type": "long"},
-			"total_agent_executions":      map[string]interface{}{"type": "long"},
-			"total_agent_tokens":          map[string]interface{}{"type": "long"},
-			"total_ai_executions":         map[string]interface{}{"type": "long"},
-			"total_app_executions_failed": map[string]interface{}{"type": "long"},
-		},
-	},
-	"datastore_ngram": {
-		"properties": map[string]interface{}{
-			"key":    map[string]interface{}{"type": "keyword"},
-			"org_id": map[string]interface{}{"type": "keyword"},
-			"amount": map[string]interface{}{"type": "long"},
-		},
-	},
-	"workflow": {
-		"properties": map[string]interface{}{
-			"id":           map[string]interface{}{"type": "keyword"},
-			"org_id":       map[string]interface{}{"type": "keyword"},
-			"created":      map[string]interface{}{"type": "date", "format": "epoch_second"},
-			"edited":       map[string]interface{}{"type": "date", "format": "epoch_second"},
-			"last_runtime": map[string]interface{}{"type": "long"},
-		},
-	},
-	"workflowapp": {
-		"properties": map[string]interface{}{
-			"app_id":      map[string]interface{}{"type": "keyword"},
-			"app_version": map[string]interface{}{"type": "keyword"},
-			"generated":   map[string]interface{}{"type": "boolean"},
-			"small_image": map[string]interface{}{"type": "keyword", "index": false, "doc_values": false},
-			"large_image": map[string]interface{}{"type": "keyword", "index": false, "doc_values": false},
-		},
-	},
-	"org_cache": {
-		"properties": map[string]interface{}{
-			"key":    map[string]interface{}{"type": "keyword"},
-			"org_id": map[string]interface{}{"type": "keyword"},
-		},
-	},
-}
-
 func SetOrgStatistics(ctx context.Context, stats ExecutionInfo, id string) error {
 	nameKey := "org_statistics"
 
@@ -20194,840 +20020,6 @@ func HealthCheckHandler(resp http.ResponseWriter, request *http.Request) {
 	//fmt.Fprint(res, "OK")
 }
 
-// resolveAppendIndexCreationTarget decides which concrete backing index the
-// create-loop in InitOpensearchIndexes should target for a given append
-// (rollover) base index: either a brand new "-000001" generation (if none
-// exists yet) or the highest EXISTING generation (if the index has already
-// been created/rolled/collapsed before). existingIndices is the full list
-// of real index names currently in the cluster (from getOpensearchIndices).
-func resolveAppendIndexCreationTarget(existingIndices []string, index string) (target string, alreadyExists bool) {
-	prefix := index + "-"
-	highestGen := -1
-	highestName := ""
-
-	for _, name := range existingIndices {
-		if !strings.HasPrefix(name, prefix) {
-			continue
-		}
-
-		gen := getOpensearchGeneration(name)
-		if gen <= 0 {
-			continue
-		}
-		if gen > highestGen {
-			highestGen = gen
-			highestName = name
-		}
-	}
-
-	if highestName == "" {
-		return fmt.Sprintf("%s-000001", index), false
-	}
-
-	return highestName, true
-}
-
-func InitOpensearchIndexes() {
-	if project.DbType != "opensearch" {
-		return
-	}
-
-	if os.Getenv("SHUFFLE_SKIP_OPENSEARCH_INDEX_INIT") == "true" {
-		return
-	}
-
-	// Check if the "workflowexecution" index exists and configuring rollovers if possible
-	log.Printf("[INFO] Configuring Opensearch indexes for scaling")
-
-	ctx := context.Background()
-	opensearchUrl := strings.TrimRight(os.Getenv("SHUFFLE_OPENSEARCH_URL"), "/")
-	if len(opensearchUrl) == 0 {
-		opensearchUrl = "https://shuffle-opensearch:9200"
-	}
-
-	relevantScaleIndexes := []string{}
-	for _, baseIndex := range GetOpensearchBaseIndexes() {
-		relevantScaleIndexes = append(relevantScaleIndexes, GetESIndexPrefix(baseIndex))
-	}
-
-	// Only append-heavy stores get rollover. Stateful keyed stores stay on a
-	// single backing index (rollover there splits _id across generations and
-	// breaks single-document reads, e.g. org_statistics).
-	appendIndexes := []string{}
-	for _, baseIndex := range GetOpensearchRolloverIndexes() {
-		appendIndexes = append(appendIndexes, strings.ToLower(GetESIndexPrefix(baseIndex)))
-	}
-
-	singleIndexes := []string{}
-	for _, index := range relevantScaleIndexes {
-		index = strings.ToLower(index)
-		if !ArrayContains(appendIndexes, index) {
-			singleIndexes = append(singleIndexes, index)
-		}
-	}
-
-	customConfig := os.Getenv("OPENSEARCH_INDEX_CONFIG")
-	if len(customConfig) > 0 {
-		checkValidJson := map[string]interface{}{}
-		if err := json.Unmarshal([]byte(customConfig), &checkValidJson); err != nil {
-			log.Printf("[ERROR] Invalid JSON in OPENSEARCH_INDEX_CONFIG: %s", err)
-			customConfig = ""
-		}
-
-		log.Printf("[DEBUG] Using custom index config for relevant scale indexes: %s", customConfig)
-	}
-
-	customRollover := os.Getenv("OPENSEARCH_INDEX_ROLLOVER")
-	if len(customRollover) > 0 {
-		checkValidJson := map[string]interface{}{}
-		if err := json.Unmarshal([]byte(customRollover), &checkValidJson); err != nil {
-			log.Printf("[ERROR] Invalid JSON in OPENSEARCH_INDEX_ROLLOVER: %s", err)
-			customRollover = ""
-		}
-
-		log.Printf("[DEBUG] Using custom rollover config for relevant scale indexes: %s", customRollover)
-	}
-
-	rolloverConfig := []byte(fmt.Sprintf(`{
-		"conditions": {
-			"max_age": "90d",
-			"max_size": "40gb",
-			"max_docs": 1000000
-		}
-	}`))
-
-	if len(customRollover) > 0 {
-		rolloverConfig = []byte(customRollover)
-	}
-
-	ismEnabled := strings.ToLower(strings.TrimSpace(os.Getenv("OPENSEARCH_USE_ISM_ROLLOVER"))) != "false"
-	ismPolicyName := strings.TrimSpace(os.Getenv("OPENSEARCH_ISM_POLICY_NAME"))
-	if ismPolicyName == "" {
-		ismPolicyName = "shuffle-rollover"
-	}
-
-	ismReady := false
-	if ismEnabled {
-		for _, baseIndex := range GetOpensearchRolloverIndexes() {
-			alias := strings.ToLower(GetESIndexPrefix(baseIndex))
-			retention := getOpensearchRetentionDays(baseIndex)
-			ready, err := ensureOpensearchISMRolloverPolicy(ctx, opensearchUrl, alias, rolloverConfig, retention, ismPolicyName)
-			if err != nil {
-				log.Printf("[WARNING] Failed ensuring ISM rollover policy '%s': %s", ismPolicyName, err)
-				continue
-			}
-			if ready {
-				ismReady = true
-			}
-		}
-	}
-
-	if fixResult, fixErr := FixOpensearchIndexPrefix(ctx); fixErr != nil {
-		log.Printf("[WARNING] Prefix repair before init failed: %s", fixErr)
-	} else if !fixResult.Success {
-		log.Printf("[WARNING] Prefix repair before init completed with verification warnings: %s", fixResult.Reason)
-	} else {
-		log.Printf("[INFO] Prefix repair before init: expected aliases=%d found=%d", fixResult.ExpectedAliases, fixResult.FoundAliases)
-	}
-
-	if len(customConfig) == 0 {
-		ensureOpensearchMappingTemplates(ctx, opensearchUrl)
-	}
-
-	existingOpensearchIndices, existingIndicesErr := getOpensearchIndices(project.Es, opensearchUrl)
-	if existingIndicesErr != nil {
-		log.Printf("[WARNING] Failed listing existing OpenSearch indices before create-loop (falling back to blind -000001 creation for all indexes): %s", existingIndicesErr)
-		existingOpensearchIndices = []string{}
-	}
-
-	for _, index := range relevantScaleIndexes {
-		indexConfig := []byte(fmt.Sprintf(`{
-			"aliases": {
-				"%s": {
-					"is_write_index": true
-				}
-			},
-			"settings": {
-				"number_of_shards": 3,
-				"number_of_replicas": 1,
-				"refresh_interval": "30s"
-			},
-			"mappings": {
-				"dynamic_templates": [
-					{
-						"strings_as_keywords": {
-							"match_mapping_type": "string",
-							"mapping": {
-								"type": "keyword",
-								"ignore_above": 256
-							}
-						}
-					}
-				]
-			}
-		}`, index))
-
-		if len(customConfig) > 0 {
-			indexConfig = []byte(customConfig)
-
-			// Check if alias is in the index or not, otherwise inject it
-			unmarshalled := map[string]interface{}{}
-			if err := json.Unmarshal(indexConfig, &unmarshalled); err != nil {
-				log.Printf("[ERROR] Invalid JSON in OPENSEARCH_INDEX_CONFIG (2): %s", err)
-			} else {
-				if _, ok := unmarshalled["aliases"]; !ok {
-					// Inject it
-					aliasPart := map[string]interface{}{
-						index: map[string]bool{
-							"is_write_index": true,
-						},
-					}
-					unmarshalled["aliases"] = aliasPart
-					newConfig, err := json.Marshal(unmarshalled)
-					if err != nil {
-						log.Printf("[ERROR] Invalid JSON in OPENSEARCH_INDEX_CONFIG (3): %s", err)
-					} else {
-						indexConfig = newConfig
-						log.Printf("[INFO] Injected alias into OPENSEARCH_INDEX_CONFIG for index %s", index)
-					}
-				}
-			}
-
-		}
-
-		index = strings.ToLower(index)
-		isAppend := ArrayContains(appendIndexes, index)
-		if len(customConfig) == 0 {
-			indexConfig = applyOpensearchCoreMappings(indexConfig, index)
-		}
-		initialIndexName, alreadyExists := resolveAppendIndexCreationTarget(existingOpensearchIndices, index)
-		if isAppend {
-			indexConfig = ensureOpensearchIndexRolloverAlias(indexConfig, index)
-		}
-		// Directly try to force create it. Opensearch throws a 400 if it fails.
-
-		var resp *opensearchapi.IndicesCreateResp
-		var err error
-		if alreadyExists {
-			log.Printf("[INFO] Index %s already exists at generation %s - skipping creation, ensuring ISM/rollover on existing index", index, initialIndexName)
-		} else {
-			resp, err = project.Es.Indices.Create(ctx, opensearchapi.IndicesCreateReq{
-				Index: initialIndexName,
-				Body:  bytes.NewReader(indexConfig),
-			})
-
-			res := resp.Inspect().Response
-			defer res.Body.Close()
-			if err != nil {
-				if !strings.Contains(fmt.Sprintf("%s", err), "serverless mode") && !strings.Contains(fmt.Sprintf("%s", err), "resource_already_exists_exception") {
-					log.Printf("[WARNING] Error creating index %s: %s", index, err)
-				}
-
-				// Make sure if the resource exist it is part of correct alias
-				if strings.Contains(fmt.Sprintf("%s", err), "resource_already_exists_exception") {
-					body := fmt.Sprintf(`{
-					  "actions": [
-					    {
-					      "add": {
-					        "index": "%s",
-					        "alias": "%s",
-					        "is_write_index": true
-					      }
-					    }
-					  ]
-					}`, initialIndexName, index)
-
-					aliasResp, aerr := project.Es.Aliases(ctx, opensearchapi.AliasesReq{
-						Body: strings.NewReader(body),
-					})
-					if aerr != nil {
-						log.Printf("[WARNING] Failed to ensure alias %s for index %s: %s", index, initialIndexName, aerr)
-						return
-					}
-
-					res := aliasResp.Inspect().Response
-					defer res.Body.Close()
-
-					if res.StatusCode >= 300 {
-						log.Printf("[WARNING] Alias enforcement failed: %s", res.String())
-						return
-					}
-				}
-			} else {
-				if res.IsError() {
-					if !strings.Contains(res.String(), "resource_already_exists_exception") {
-						log.Printf("[DEBUG] Error creating index %s with custom config: %s", index, res.String())
-					}
-
-				} else {
-					log.Printf("[DEBUG] Successfully created index %s with custom config", index)
-				}
-			}
-		}
-
-		// Non-append indexes stay on a single backing index - no rollover/ISM.
-		if !isAppend {
-			continue
-		}
-
-		if ismReady {
-			if err := ensureOpensearchIndexRolloverAliasSetting(ctx, opensearchUrl, initialIndexName, index); err != nil {
-				log.Printf("[WARNING] Failed ensuring rollover_alias on index %s: %s", initialIndexName, err)
-			}
-
-			policyID := fmt.Sprintf("%s-%s", ismPolicyName, index)
-			if err := ensureOpensearchIndexISMPolicy(ctx, opensearchUrl, initialIndexName, policyID); err != nil {
-				log.Printf("[WARNING] Failed attaching ISM policy '%s' to %s: %s", policyID, initialIndexName, err)
-			}
-
-			continue
-		}
-
-		rolloverResp, err := project.Es.Indices.Rollover(ctx, opensearchapi.IndicesRolloverReq{
-			Alias: index,
-			Body:  bytes.NewReader(rolloverConfig),
-		})
-
-		if err != nil {
-			if !strings.Contains(fmt.Sprintf("%s", err), "serverless mode") && !strings.Contains(fmt.Sprintf("%s", err), "status: 404") {
-				log.Printf("[WARNING] Problem during rollover config for %s: %s", index, err)
-			}
-
-			continue
-		}
-
-		rolloverRes := rolloverResp.Inspect().Response
-		defer rolloverRes.Body.Close()
-		if rolloverRes.IsError() {
-			log.Printf("[ERROR] Rollover config failed for %s: %s", index, rolloverRes.String())
-		} else {
-			log.Printf("[INFO] Rollover executed successfully for %s", index)
-		}
-
-	}
-
-	// Migrate existing deployments that rolled stateful indexes in the past:
-	// collapse all generations of each single index into its newest backing
-	// index and detach ISM so it never rolls again. Idempotent.
-	for _, singleIndex := range singleIndexes {
-		if err := collapseSingleIndexAliases(ctx, opensearchUrl, singleIndex); err != nil {
-			log.Printf("[WARNING] Failed collapsing single index %s: %s", singleIndex, err)
-		}
-	}
-
-	// Apply mapping migrations to existing single/keyed indexes when the live
-	// mapping has drifted from opensearchCoreMappings. Skipped when a custom
-	// OPENSEARCH_INDEX_CONFIG is set (the operator owns those mappings).
-	if len(customConfig) == 0 {
-		for _, singleIndex := range singleIndexes {
-			if err := migrateOpensearchSingleIndex(ctx, opensearchUrl, singleIndex); err != nil {
-				log.Printf("[WARNING] Failed migrating mapping for single index %s: %s", singleIndex, err)
-			}
-		}
-	}
-
-	if fixResult, fixErr := FixOpensearchIndexPrefix(ctx); fixErr != nil {
-		log.Printf("[WARNING] Alias verification after init failed: %s", fixErr)
-	} else if !fixResult.Success {
-		log.Printf("[WARNING] Alias verification after init completed with warnings: %s", fixResult.Reason)
-	} else {
-		log.Printf("[INFO] Alias verification after init passed: expected aliases=%d found=%d", fixResult.ExpectedAliases, fixResult.FoundAliases)
-	}
-
-}
-
-func ensureOpensearchIndexRolloverAlias(indexConfig []byte, alias string) []byte {
-	unmarshalled := map[string]interface{}{}
-	if err := json.Unmarshal(indexConfig, &unmarshalled); err != nil {
-		return indexConfig
-	}
-
-	settings, ok := unmarshalled["settings"].(map[string]interface{})
-	if !ok || settings == nil {
-		settings = map[string]interface{}{}
-	}
-
-	settings["plugins.index_state_management.rollover_alias"] = alias
-	unmarshalled["settings"] = settings
-
-	updated, err := json.Marshal(unmarshalled)
-	if err != nil {
-		return indexConfig
-	}
-
-	return updated
-}
-
-// applyOpensearchCoreMappings injects the pragmatic core field mappings for a
-// base index into a fresh index-create body. Only used when no custom
-// OPENSEARCH_INDEX_CONFIG is set, and ignored by OpenSearch for indices that
-// already exist (mappings are immutable).
-func applyOpensearchCoreMappings(indexConfig []byte, index string) []byte {
-	key := strings.TrimPrefix(strings.ToLower(index), strings.ToLower(GetESIndexPrefix("")))
-	properties, ok := opensearchCoreMappings[key]
-	if !ok {
-		return indexConfig
-	}
-
-	unmarshalled := map[string]interface{}{}
-	if err := json.Unmarshal(indexConfig, &unmarshalled); err != nil {
-		return indexConfig
-	}
-
-	mappings, _ := unmarshalled["mappings"].(map[string]interface{})
-	if mappings == nil {
-		mappings = map[string]interface{}{}
-	}
-
-	if _, exists := mappings["properties"]; !exists {
-		mappings["properties"] = properties["properties"]
-	}
-	unmarshalled["mappings"] = mappings
-
-	updated, err := json.Marshal(unmarshalled)
-	if err != nil {
-		return indexConfig
-	}
-
-	return updated
-}
-
-// opensearchMappingsFor builds the mappings section (dynamic string->keyword
-// template plus the curated core field mappings) for a base index. Mirrors what
-// a freshly created index resolves to, so the migration check stays consistent
-// with fresh creation.
-func opensearchMappingsFor(baseIndex string) map[string]interface{} {
-	mappings := map[string]interface{}{
-		"dynamic_templates": []map[string]interface{}{
-			{
-				"strings_as_keywords": map[string]interface{}{
-					"match_mapping_type": "string",
-					"mapping":            map[string]interface{}{"type": "keyword", "ignore_above": 256},
-				},
-			},
-		},
-	}
-
-	if props, ok := opensearchCoreMappings[baseIndex]["properties"]; ok {
-		mappings["properties"] = props
-	}
-
-	return mappings
-}
-
-// opensearchMappingsDiffer reports whether the live index mapping (its
-// "properties" subtree) is missing any curated field, or has a field whose
-// type/format/indexability no longer matches the desired core mappings. Only
-// fields we explicitly map are compared - the many auto-mapped string fields
-// are ignored.
-func opensearchMappingsDiffer(baseIndex string, actualProps map[string]interface{}) bool {
-	desired, ok := opensearchCoreMappings[baseIndex]["properties"].(map[string]interface{})
-	if !ok {
-		return false
-	}
-
-	for name, dv := range desired {
-		desiredProp, _ := dv.(map[string]interface{})
-		actualProp, exists := actualProps[name].(map[string]interface{})
-		if !exists {
-			return true
-		}
-
-		if fmt.Sprint(desiredProp["type"]) != fmt.Sprint(actualProp["type"]) {
-			return true
-		}
-
-		if df, ok := desiredProp["format"]; ok && fmt.Sprint(actualProp["format"]) != fmt.Sprint(df) {
-			return true
-		}
-
-		if fmt.Sprint(desiredProp["index"]) == "false" && fmt.Sprint(actualProp["index"]) != "false" {
-			return true
-		}
-	}
-
-	return false
-}
-
-// getOpensearchIndexProperties returns the "properties" subtree of an index's
-// live mappings.
-func getOpensearchIndexProperties(foundClient opensearchapi.Client, opensearchUrl, indexName string) (map[string]interface{}, error) {
-	mappingReq, err := http.NewRequest("GET", fmt.Sprintf("%s/%s/_mapping", opensearchUrl, indexName), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	mappingResp, err := foundClient.Client.Transport.Perform(mappingReq)
-	if err != nil {
-		return nil, err
-	}
-	defer mappingResp.Body.Close()
-
-	body, err := ioutil.ReadAll(mappingResp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if mappingResp.StatusCode >= 300 {
-		return nil, fmt.Errorf("failed reading mapping for %s: %d %s", indexName, mappingResp.StatusCode, string(body))
-	}
-
-	raw := map[string]interface{}{}
-	if err := json.Unmarshal(body, &raw); err != nil {
-		return nil, err
-	}
-
-	for _, idxVal := range raw {
-		idxMap, _ := idxVal.(map[string]interface{})
-		mappings, _ := idxMap["mappings"].(map[string]interface{})
-		props, _ := mappings["properties"].(map[string]interface{})
-		return props, nil
-	}
-
-	return nil, nil
-}
-
-// createOpensearchIndexFromBody creates an index with an explicit create body.
-func createOpensearchIndexFromBody(ctx context.Context, opensearchUrl, indexName string, body []byte) error {
-	req, err := http.NewRequestWithContext(ctx, "PUT", fmt.Sprintf("%s/%s", opensearchUrl, indexName), bytes.NewReader(body))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := project.Es.Client.Transport.Perform(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	respBody, _ := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode >= 300 {
-		return fmt.Errorf("failed creating index %s: %d %s", indexName, resp.StatusCode, string(respBody))
-	}
-
-	return nil
-}
-
-// migrateOpensearchSingleIndex re-creates a single (keyed) index with the
-// current core mappings when its live mapping has drifted. It reindexes the
-// existing backing index into a fresh generation, atomically swaps the alias to
-// the new generation (so the alias never points at two indices), then drops the
-// old one. Safe to run repeatedly - it no-ops once mappings match.
-func migrateOpensearchSingleIndex(ctx context.Context, opensearchUrl, baseIndex string) error {
-	foundClient := project.Es
-	allIndices, err := getOpensearchIndices(foundClient, opensearchUrl)
-	if err != nil {
-		return err
-	}
-
-	generations := []string{}
-	for _, idx := range allIndices {
-		if idx == baseIndex || strings.HasPrefix(idx, baseIndex+"-") {
-			generations = append(generations, idx)
-		}
-	}
-	if len(generations) == 0 {
-		return nil
-	}
-
-	sort.Slice(generations, func(i, j int) bool {
-		return getOpensearchGeneration(generations[i]) > getOpensearchGeneration(generations[j])
-	})
-
-	// collapseSingleIndexAliases runs just before this; if multiple generations
-	// remain, defer to it rather than racing a partial collapse.
-	if len(generations) > 1 {
-		return nil
-	}
-
-	src := generations[0]
-	actualProps, err := getOpensearchIndexProperties(foundClient, opensearchUrl, src)
-	if err != nil {
-		return err
-	}
-	if !opensearchMappingsDiffer(baseIndex, actualProps) {
-		return nil
-	}
-
-	nextGen := getOpensearchGeneration(src) + 1
-	dest := fmt.Sprintf("%s-%06d", baseIndex, nextGen)
-
-	body := map[string]interface{}{
-		"settings": map[string]interface{}{
-			"number_of_shards":   3,
-			"number_of_replicas": 1,
-			"refresh_interval":   "30s",
-		},
-		"mappings": opensearchMappingsFor(baseIndex),
-	}
-	bodyJSON, err := json.Marshal(body)
-	if err != nil {
-		return err
-	}
-
-	if err := createOpensearchIndexFromBody(ctx, opensearchUrl, dest, bodyJSON); err != nil {
-		return err
-	}
-	if err := reindexOpensearchIndex(ctx, opensearchUrl, src, dest); err != nil {
-		return err
-	}
-
-	// atomically move the write alias from the old generation to the new one
-	write := true
-	actions := []OpensearchAliasAction{
-		{Remove: &OpensearchAliasActionTarget{Index: src, Alias: baseIndex}},
-		{Add: &OpensearchAliasActionTarget{Index: dest, Alias: baseIndex, IsWriteIndex: &write}},
-	}
-	if err := updateOpensearchAliases(foundClient, opensearchUrl, actions); err != nil {
-		return err
-	}
-
-	if err := deleteOpensearchIndex(foundClient, opensearchUrl, src); err != nil {
-		return err
-	}
-
-	log.Printf("[INFO] Migrated mapping for %s: %s -> %s", baseIndex, src, dest)
-	return nil
-}
-
-// ensureOpensearchMappingTemplates registers an index mapping template per
-// append/rollover base index so every future rollover generation is created
-// with the current core mappings (existing generations are left untouched).
-func ensureOpensearchMappingTemplates(ctx context.Context, opensearchUrl string) {
-	for _, baseIndex := range GetOpensearchRolloverIndexes() {
-		alias := strings.ToLower(GetESIndexPrefix(baseIndex))
-
-		body := map[string]interface{}{
-			"index_patterns": []string{alias + "-*"},
-			"template": map[string]interface{}{
-				"mappings": opensearchMappingsFor(baseIndex),
-			},
-			"priority": 100,
-		}
-		bodyJSON, err := json.Marshal(body)
-		if err != nil {
-			log.Printf("[WARNING] Failed building mapping template for %s: %s", alias, err)
-			continue
-		}
-
-		templateName := fmt.Sprintf("shuffle-%s-mapping", baseIndex)
-		req, err := http.NewRequestWithContext(ctx, "PUT", fmt.Sprintf("%s/_index_template/%s", opensearchUrl, templateName), bytes.NewReader(bodyJSON))
-		if err != nil {
-			log.Printf("[WARNING] Failed to create mapping template request for %s: %s", alias, err)
-			continue
-		}
-		req.Header.Set("Content-Type", "application/json")
-
-		resp, err := project.Es.Client.Transport.Perform(req)
-		if err != nil {
-			log.Printf("[WARNING] Failed to register mapping template for %s: %s", alias, err)
-			continue
-		}
-		_ = resp.Body.Close()
-		if resp.StatusCode >= 300 {
-			log.Printf("[WARNING] Failed to register mapping template for %s: status %d", alias, resp.StatusCode)
-		}
-	}
-}
-
-// collapseSingleIndexAliases migrates a stateful (non-append) index that may
-// have rolled over in previous versions to a single backing index: it merges
-// every older generation into the newest (newest document wins per _id), drops
-// the older generations, and detaches rollover so the index stays single. Safe
-// to run repeatedly.
-func collapseSingleIndexAliases(ctx context.Context, opensearchUrl, fullIndex string) error {
-	foundClient := project.Es
-	allIndices, err := getOpensearchIndices(foundClient, opensearchUrl)
-	if err != nil {
-		return err
-	}
-
-	generations := []string{}
-	for _, idx := range allIndices {
-		if idx == fullIndex || strings.HasPrefix(idx, fullIndex+"-") {
-			generations = append(generations, idx)
-		}
-	}
-
-	if len(generations) == 1 {
-		// A single surviving index - just make sure it can't roll over.
-		return detachOpensearchRollover(ctx, opensearchUrl, generations[0])
-	}
-	if len(generations) == 0 {
-		return nil
-	}
-
-	sort.Slice(generations, func(i, j int) bool {
-		return getOpensearchGeneration(generations[i]) > getOpensearchGeneration(generations[j])
-	})
-	writeGen := generations[0]
-	olderGens := generations[1:]
-
-	// Merge older generations into the newest (the surviving write target).
-	// reindexOpensearchIndex uses op_type:create, so a source _id that already
-	// exists in the newest generation is skipped (conflicts=proceed) and the
-	// newest copy wins; _ids that only live in older generations are copied
-	// across. Iteration order does not matter because the destination always
-	// wins on collision.
-	for _, older := range olderGens {
-		if err := reindexOpensearchIndex(ctx, opensearchUrl, older, writeGen); err != nil {
-			return err
-		}
-	}
-
-	actions := []OpensearchAliasAction{}
-	for _, older := range olderGens {
-		actions = append(actions, OpensearchAliasAction{
-			Remove: &OpensearchAliasActionTarget{Index: older, Alias: fullIndex},
-		})
-	}
-	if err := updateOpensearchAliases(foundClient, opensearchUrl, actions); err != nil {
-		return err
-	}
-
-	for _, older := range olderGens {
-		if err := deleteOpensearchIndex(foundClient, opensearchUrl, older); err != nil {
-			return err
-		}
-	}
-
-	return detachOpensearchRollover(ctx, opensearchUrl, writeGen)
-}
-
-// opensearchReindexBody builds the _reindex request body. op_type must live on
-// the dest object (per the Reindex API): "create" keeps an existing destination
-// document and skips the source copy on _id collision (newest-wins when merging
-// generations). Placing it at the top level would be silently ignored, so this
-// is kept pure and unit-tested.
-func opensearchReindexBody(source, dest string) map[string]interface{} {
-	return map[string]interface{}{
-		"source": map[string]interface{}{"index": source},
-		"dest": map[string]interface{}{
-			"index":   dest,
-			"op_type": "create",
-		},
-		"conflicts": "proceed",
-	}
-}
-
-func reindexOpensearchIndex(ctx context.Context, opensearchUrl, source, dest string) error {
-	body := opensearchReindexBody(source, dest)
-
-	data, err := json.Marshal(body)
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequestWithContext(ctx, "POST", fmt.Sprintf("%s/_reindex", opensearchUrl), bytes.NewReader(data))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := project.Es.Client.Transport.Perform(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	respBody, _ := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode >= 300 {
-		return fmt.Errorf("reindex %s->%s failed: %d %s", source, dest, resp.StatusCode, string(respBody))
-	}
-
-	return nil
-}
-
-// detachOpensearchRollover removes the ISM rollover policy and clears the
-// rollover_alias index setting so a single (non-append) index never rolls over.
-func detachOpensearchRollover(ctx context.Context, opensearchUrl, indexName string) error {
-	// Remove the ISM rollover policy, if any. Missing policy / missing plugin
-	// (4xx) is fine - clearing the rollover_alias setting below is what truly
-	// stops rollover.
-	req, err := http.NewRequestWithContext(ctx, "POST", fmt.Sprintf("%s/_plugins/_ism/remove/%s", opensearchUrl, indexName), strings.NewReader("{}"))
-	if err == nil {
-		req.Header.Set("Content-Type", "application/json")
-		if removeResp, performErr := project.Es.Client.Transport.Perform(req); performErr == nil {
-			_ = removeResp.Body.Close()
-		}
-	}
-
-	settingsBody := map[string]interface{}{
-		"index": map[string]interface{}{
-			"plugins.index_state_management.rollover_alias": nil,
-		},
-	}
-
-	bodyData, marshalErr := json.Marshal(settingsBody)
-	if marshalErr != nil {
-		return marshalErr
-	}
-
-	settingsReq, settingsErr := http.NewRequestWithContext(ctx, "PUT", fmt.Sprintf("%s/%s/_settings", opensearchUrl, indexName), bytes.NewReader(bodyData))
-	if settingsErr != nil {
-		return settingsErr
-	}
-	settingsReq.Header.Set("Content-Type", "application/json")
-
-	settingsResp, settingsPerformErr := project.Es.Client.Transport.Perform(settingsReq)
-	if settingsPerformErr != nil {
-		return settingsPerformErr
-	}
-	defer settingsResp.Body.Close()
-
-	respBody, _ := ioutil.ReadAll(settingsResp.Body)
-	if settingsResp.StatusCode >= 300 && settingsResp.StatusCode != 404 {
-		return fmt.Errorf("clear rollover_alias on %s failed: %d %s", indexName, settingsResp.StatusCode, string(respBody))
-	}
-
-	return nil
-}
-
-func getOpensearchISMRolloverConditions(rolloverConfig []byte) map[string]interface{} {
-	defaultConditions := map[string]interface{}{
-		"min_index_age": "90d",
-		"min_size":      "40gb",
-		"min_doc_count": 1000000,
-	}
-
-	parsed := struct {
-		Conditions map[string]interface{} `json:"conditions"`
-	}{}
-
-	if err := json.Unmarshal(rolloverConfig, &parsed); err != nil {
-		return defaultConditions
-	}
-
-	if len(parsed.Conditions) == 0 {
-		return defaultConditions
-	}
-
-	conditions := map[string]interface{}{}
-	if value, ok := parsed.Conditions["min_index_age"]; ok {
-		conditions["min_index_age"] = value
-	} else if value, ok := parsed.Conditions["max_age"]; ok {
-		conditions["min_index_age"] = value
-	}
-
-	if value, ok := parsed.Conditions["min_size"]; ok {
-		conditions["min_size"] = value
-	} else if value, ok := parsed.Conditions["max_size"]; ok {
-		conditions["min_size"] = value
-	}
-
-	if value, ok := parsed.Conditions["min_doc_count"]; ok {
-		conditions["min_doc_count"] = value
-	} else if value, ok := parsed.Conditions["max_docs"]; ok {
-		conditions["min_doc_count"] = value
-	}
-
-	if len(conditions) == 0 {
-		return defaultConditions
-	}
-
-	return conditions
-}
-
 // StartExecutionLifecycleJobs starts the background jobs that keep
 // workflowexecution_live bounded.
 func StartExecutionLifecycleJobs(ctx context.Context) {
@@ -21048,312 +20040,6 @@ func StartExecutionLifecycleJobs(ctx context.Context) {
 	}
 
 	go StartNotificationRetentionSweeper(context.Background())
-}
-
-func getOpensearchRetentionDays(baseIndex string) string {
-	defaults := map[string]string{
-		"shuffle_logs":      "90d",
-		"workflowexecution": "365d",
-	}
-
-	value := defaults[baseIndex]
-	if value == "" {
-		return ""
-	}
-
-	custom := strings.TrimSpace(os.Getenv("OPENSEARCH_INDEX_RETENTION_DAYS"))
-	if custom == "" {
-		return value
-	}
-
-	parsed := map[string]interface{}{}
-	if err := json.Unmarshal([]byte(custom), &parsed); err != nil {
-		log.Printf("[WARNING] Invalid JSON in OPENSEARCH_INDEX_RETENTION_DAYS: %s", err)
-		return value
-	}
-
-	raw, ok := parsed[baseIndex]
-	if !ok {
-		return value
-	}
-	if days, ok := raw.(float64); ok {
-		return fmt.Sprintf("%dd", int(days))
-	}
-	if str, ok := raw.(string); ok {
-		return str
-	}
-
-	return value
-}
-
-// existingOpensearchISMPolicy holds the parts of a GET
-// /_plugins/_ism/policies/<id> response needed to decide whether the policy
-// needs updating, and (if so) to perform a conflict-safe PUT.
-type existingOpensearchISMPolicy struct {
-	SeqNo         int64                  `json:"_seq_no"`
-	PrimaryTerm   int64                  `json:"_primary_term"`
-	RawConditions map[string]interface{} // hot state's rollover conditions
-	RawRetention  string                 // delete transition's min_index_age, if any
-}
-
-func getExistingOpensearchISMPolicy(ctx context.Context, opensearchUrl, policyID string) (*existingOpensearchISMPolicy, bool, error) {
-	req, err := http.NewRequestWithContext(ctx, "GET", fmt.Sprintf("%s/_plugins/_ism/policies/%s", opensearchUrl, policyID), nil)
-	if err != nil {
-		return nil, false, err
-	}
-
-	resp, err := project.Es.Client.Transport.Perform(req)
-	if err != nil {
-		return nil, false, err
-	}
-	defer resp.Body.Close()
-
-	body, _ := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode >= 300 {
-		if resp.StatusCode == 404 || resp.StatusCode == 400 {
-			if strings.Contains(strings.ToLower(string(body)), "_plugins/_ism") || strings.Contains(strings.ToLower(string(body)), "no handler found") {
-				// ISM plugin isn't installed at all.
-				return nil, false, fmt.Errorf("ism plugin not available")
-			}
-		}
-
-		if resp.StatusCode == 404 {
-			// Genuinely doesn't exist yet - needs to be created.
-			return nil, false, nil
-		}
-
-		return nil, false, fmt.Errorf("status: %d, body: %s", resp.StatusCode, string(body))
-	}
-
-	parsed := struct {
-		SeqNo       int64 `json:"_seq_no"`
-		PrimaryTerm int64 `json:"_primary_term"`
-		Policy      struct {
-			States []struct {
-				Name    string `json:"name"`
-				Actions []struct {
-					Rollover map[string]interface{} `json:"rollover"`
-				} `json:"actions"`
-				Transitions []struct {
-					Conditions struct {
-						MinIndexAge string `json:"min_index_age"`
-					} `json:"conditions"`
-				} `json:"transitions"`
-			} `json:"states"`
-		} `json:"policy"`
-	}{}
-
-	if err := json.Unmarshal(body, &parsed); err != nil {
-		return nil, false, err
-	}
-
-	existing := &existingOpensearchISMPolicy{
-		SeqNo:       parsed.SeqNo,
-		PrimaryTerm: parsed.PrimaryTerm,
-	}
-	for _, state := range parsed.Policy.States {
-		if state.Name != "hot" {
-			continue
-		}
-		if len(state.Actions) > 0 {
-			existing.RawConditions = state.Actions[0].Rollover
-		}
-		if len(state.Transitions) > 0 {
-			existing.RawRetention = state.Transitions[0].Conditions.MinIndexAge
-		}
-	}
-
-	return existing, true, nil
-}
-
-func ensureOpensearchISMRolloverPolicy(ctx context.Context, opensearchUrl, alias string, rolloverConfig []byte, retentionAge, policyName string) (bool, error) {
-	conditions := getOpensearchISMRolloverConditions(rolloverConfig)
-	policyID := fmt.Sprintf("%s-%s", policyName, alias)
-
-	states := []map[string]interface{}{
-		{
-			"name":        "hot",
-			"actions":     []map[string]interface{}{{"rollover": conditions}},
-			"transitions": []interface{}{},
-		},
-	}
-
-	if retentionAge != "" {
-		states[0]["transitions"] = []map[string]interface{}{
-			{
-				"state_name": "delete",
-				"conditions": map[string]interface{}{"min_index_age": retentionAge},
-			},
-		}
-		states = append(states, map[string]interface{}{
-			"name":        "delete",
-			"actions":     []map[string]interface{}{{"delete": map[string]interface{}{}}},
-			"transitions": []interface{}{},
-		})
-	}
-
-	policyBody := map[string]interface{}{
-		"policy": map[string]interface{}{
-			"description":   "Shuffle rollover + retention policy",
-			"default_state": "hot",
-			"states":        states,
-			"ism_template": []map[string]interface{}{
-				{
-					"index_patterns": []string{fmt.Sprintf("%s-*", alias)},
-					"priority":       100,
-				},
-			},
-		},
-	}
-
-	policyData, err := json.Marshal(policyBody)
-	if err != nil {
-		return false, err
-	}
-
-	// Check whether the policy already exists, and if so, whether its
-	// rollover conditions/retention already match what we'd write - this
-	// lets us both (a) avoid a needless PUT (and its 409) when nothing
-	// changed, and (b) actually apply changes to OPENSEARCH_INDEX_ROLLOVER /
-	// OPENSEARCH_INDEX_RETENTION_DAYS on restart when something did change,
-	// which a blind "create-only" PUT can never do once the policy exists.
-	existing, found, err := getExistingOpensearchISMPolicy(ctx, opensearchUrl, policyID)
-	if err != nil {
-		if err.Error() == "ism plugin not available" {
-			log.Printf("[INFO] ISM plugin not available. Falling back to direct rollover")
-			return false, nil
-		}
-		return false, err
-	}
-
-	putUrl := fmt.Sprintf("%s/_plugins/_ism/policies/%s", opensearchUrl, policyID)
-	if found {
-		// Compare only the specific rollover condition keys Shuffle manages,
-		// not a full deep-equal of the stored object: OpenSearch enriches
-		// the stored rollover conditions with its own extra fields we never
-		// set (e.g. "copy_alias": false), so a full-map compare would never
-		// match and would cause a needless PUT (and misleading "changed -
-		// updating" log) on every single restart. %v formatting sidesteps
-		// int (our defaults) vs float64 (values decoded from OpenSearch's
-		// JSON response) type mismatches on otherwise-equal numbers.
-		conditionsMatch := true
-		for _, key := range []string{"min_index_age", "min_size", "min_doc_count"} {
-			if fmt.Sprintf("%v", existing.RawConditions[key]) != fmt.Sprintf("%v", conditions[key]) {
-				conditionsMatch = false
-				break
-			}
-		}
-		retentionMatches := existing.RawRetention == retentionAge
-		if conditionsMatch && retentionMatches {
-			log.Printf("[DEBUG] ISM rollover policy '%s' already up to date for alias %s - skipping", policyID, alias)
-			return true, nil
-		}
-
-		log.Printf("[INFO] ISM rollover policy '%s' conditions/retention changed for alias %s - updating", policyID, alias)
-		putUrl = fmt.Sprintf("%s?if_seq_no=%d&if_primary_term=%d", putUrl, existing.SeqNo, existing.PrimaryTerm)
-	}
-
-	req, err := http.NewRequestWithContext(ctx, "PUT", putUrl, bytes.NewReader(policyData))
-	if err != nil {
-		return false, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := project.Es.Client.Transport.Perform(req)
-	if err != nil {
-		return false, err
-	}
-	defer resp.Body.Close()
-
-	body, _ := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode >= 300 {
-		if resp.StatusCode == 404 || resp.StatusCode == 400 {
-			if strings.Contains(strings.ToLower(string(body)), "_plugins/_ism") || strings.Contains(strings.ToLower(string(body)), "no handler found") {
-				log.Printf("[INFO] ISM plugin not available. Falling back to direct rollover")
-				return false, nil
-			}
-		}
-
-		return false, fmt.Errorf("status: %d, body: %s", resp.StatusCode, string(body))
-	}
-
-	log.Printf("[INFO] Ensured ISM rollover policy '%s' for alias %s", policyID, alias)
-	return true, nil
-}
-
-func ensureOpensearchIndexRolloverAliasSetting(ctx context.Context, opensearchUrl, indexName, alias string) error {
-	settingsBody := map[string]interface{}{
-		"index": map[string]interface{}{
-			"plugins.index_state_management.rollover_alias": alias,
-		},
-	}
-
-	body, err := json.Marshal(settingsBody)
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequestWithContext(ctx, "PUT", fmt.Sprintf("%s/%s/_settings", opensearchUrl, indexName), bytes.NewReader(body))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := project.Es.Client.Transport.Perform(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	respBody, _ := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode >= 300 {
-		if resp.StatusCode == 404 && strings.Contains(strings.ToLower(string(respBody)), "index_not_found_exception") {
-			return nil
-		}
-
-		return fmt.Errorf("status: %d, body: %s", resp.StatusCode, string(respBody))
-	}
-
-	return nil
-}
-
-func ensureOpensearchIndexISMPolicy(ctx context.Context, opensearchUrl, indexName, policyName string) error {
-	policyBody := map[string]interface{}{
-		"policy_id": policyName,
-	}
-
-	body, err := json.Marshal(policyBody)
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequestWithContext(ctx, "POST", fmt.Sprintf("%s/_plugins/_ism/add/%s", opensearchUrl, indexName), bytes.NewReader(body))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := project.Es.Client.Transport.Perform(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	respBody, _ := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode >= 300 {
-		lowerResp := strings.ToLower(string(respBody))
-		if strings.Contains(lowerResp, "already has a policy") {
-			return nil
-		}
-
-		if resp.StatusCode == 404 && strings.Contains(lowerResp, "index_not_found_exception") {
-			return nil
-		}
-
-		return fmt.Errorf("status: %d, body: %s", resp.StatusCode, string(respBody))
-	}
-
-	return nil
 }
 
 func ListVulnerabilities(ctx context.Context, ecosystem string, inputcursor string) ([]OSVVulnerability, string, error) {
