+++ health.go
@@ -15,7 +15,6 @@ import (
 	"mime/multipart"
 	"net/http"
 	"os"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
@@ -3711,779 +3710,6 @@ func HandleRerunExecutions(resp http.ResponseWriter, request *http.Request) {
 	//log.Printf("[DEBUG] RERAN %d execution(s) in total for environment %s for org %s", total, fileId, user.ActiveOrg.Id)
 	resp.WriteHeader(200)
 	resp.Write([]byte(fmt.Sprintf(`{"success": true, "reason": "Successfully RERAN %d executions"}`, total)))
-}
-
-func FixOpensearchIndexPrefix(ctx context.Context) (OpensearchPrefixFixResult, error) {
-	result := OpensearchPrefixFixResult{}
-	if project.Environment == "cloud" {
-		result.Reason = "Opensearch prefix repair not supported in cloud"
-		return result, errors.New(result.Reason)
-	}
-
-	if project.DbType != "opensearch" {
-		result.Reason = "Opensearch is not configured"
-		return result, errors.New(result.Reason)
-	}
-
-	opensearchUrl := strings.TrimRight(os.Getenv("SHUFFLE_OPENSEARCH_URL"), "/")
-	if len(opensearchUrl) == 0 {
-		opensearchUrl = "https://shuffle-opensearch:9200"
-	}
-
-	foundClient := project.Es
-	allIndices, err := getOpensearchIndices(foundClient, opensearchUrl)
-	if err != nil {
-		return result, err
-	}
-
-	aliasInfo, err := getOpensearchAliases(foundClient, opensearchUrl)
-	if err != nil {
-		return result, err
-	}
-
-	prefix := strings.ToLower(strings.TrimSpace(os.Getenv("SHUFFLE_OPENSEARCH_INDEX_PREFIX")))
-	baseIndexes := GetOpensearchBaseIndexes()
-	expectedAliases := []string{}
-	for _, baseIndex := range baseIndexes {
-		expectedAliases = append(expectedAliases, strings.ToLower(GetESIndexPrefix(baseIndex)))
-	}
-
-	rolloverConfig := []byte(fmt.Sprintf(`{
-		"conditions": {
-			"max_age": "90d",
-			"max_size": "40gb",
-			"max_docs": 1000000
-		}
-	}`))
-
-	customRollover := os.Getenv("OPENSEARCH_INDEX_ROLLOVER")
-	if len(customRollover) > 0 {
-		checkValidJson := map[string]interface{}{}
-		if err := json.Unmarshal([]byte(customRollover), &checkValidJson); err != nil {
-			log.Printf("[ERROR] Invalid JSON in OPENSEARCH_INDEX_ROLLOVER: %s", err)
-		} else {
-			rolloverConfig = []byte(customRollover)
-		}
-	}
-
-	ismEnabled := strings.ToLower(strings.TrimSpace(os.Getenv("OPENSEARCH_USE_ISM_ROLLOVER"))) != "false"
-	ismPolicyName := strings.TrimSpace(os.Getenv("OPENSEARCH_ISM_POLICY_NAME"))
-	if ismPolicyName == "" {
-		ismPolicyName = "shuffle-rollover"
-	}
-
-	for _, baseIndex := range baseIndexes {
-		expectedAlias := strings.ToLower(GetESIndexPrefix(baseIndex))
-		doubleAlias := ""
-		if prefix != "" {
-			doubleAlias = fmt.Sprintf("%s_%s", prefix, expectedAlias)
-		}
-
-		if ArrayContains(allIndices, expectedAlias) {
-			targetIndex := fmt.Sprintf("%s-000001", expectedAlias)
-			taskID, err := handleAliasCollisionMigration(foundClient, opensearchUrl, expectedAlias, targetIndex)
-			if err != nil {
-				result.Skipped = append(result.Skipped, fmt.Sprintf("%s (collision repair failed: %s)", expectedAlias, err))
-				continue
-			}
-
-			if taskID != "" {
-				result.MigrationTasks = append(result.MigrationTasks, fmt.Sprintf("%s -> %s (task=%s)", expectedAlias, targetIndex, taskID))
-				result.Skipped = append(result.Skipped, fmt.Sprintf("%s (collision migration in progress)", expectedAlias))
-				continue
-			}
-
-			allIndices, err = getOpensearchIndices(foundClient, opensearchUrl)
-			if err != nil {
-				return result, err
-			}
-
-			aliasInfo, err = getOpensearchAliases(foundClient, opensearchUrl)
-			if err != nil {
-				return result, err
-			}
-		}
-
-		targetIndices, writeIndex := selectOpensearchAliasTargets(expectedAlias, doubleAlias, aliasInfo, allIndices)
-		if len(targetIndices) == 0 {
-			newIndex := fmt.Sprintf("%s-000001", expectedAlias)
-			if !ArrayContains(allIndices, newIndex) {
-				if err := createOpensearchIndex(foundClient, opensearchUrl, newIndex); err != nil {
-					return result, err
-				}
-				result.Created = append(result.Created, newIndex)
-				allIndices = append(allIndices, newIndex)
-			}
-
-			targetIndices = []string{newIndex}
-			writeIndex = newIndex
-		}
-
-		actions := []OpensearchAliasAction{}
-		for _, indexName := range targetIndices {
-			current, hasCurrent := aliasInfo[indexName][expectedAlias]
-			desiredWrite := indexName == writeIndex
-
-			if hasCurrent {
-				if current.IsWriteIndex != desiredWrite {
-					actions = append(actions, OpensearchAliasAction{
-						Remove: &OpensearchAliasActionTarget{Index: indexName, Alias: expectedAlias},
-					})
-					actions = append(actions, OpensearchAliasAction{
-						Add: &OpensearchAliasActionTarget{Index: indexName, Alias: expectedAlias, IsWriteIndex: &desiredWrite},
-					})
-				}
-			} else {
-				actions = append(actions, OpensearchAliasAction{
-					Add: &OpensearchAliasActionTarget{Index: indexName, Alias: expectedAlias, IsWriteIndex: &desiredWrite},
-				})
-			}
-
-			if doubleAlias != "" {
-				doubleAliasState, hasDoubleAlias := aliasInfo[indexName][doubleAlias]
-				if hasDoubleAlias && doubleAliasState.Present {
-					actions = append(actions, OpensearchAliasAction{
-						Remove: &OpensearchAliasActionTarget{Index: indexName, Alias: doubleAlias},
-					})
-				}
-			}
-		}
-
-		if len(actions) > 0 {
-			if err := updateOpensearchAliases(foundClient, opensearchUrl, actions); err != nil {
-				return result, err
-			}
-			result.AliasUpdates = append(result.AliasUpdates, fmt.Sprintf("%s -> %s", expectedAlias, writeIndex))
-		}
-
-		result.WriteIndexUpdates = append(result.WriteIndexUpdates, fmt.Sprintf("%s -> %s", expectedAlias, writeIndex))
-	}
-
-	verifiedAliasInfo, err := getOpensearchAliases(foundClient, opensearchUrl)
-	if err != nil {
-		return result, err
-	}
-
-	result.ExpectedAliases = len(expectedAliases)
-	result.FoundAliases = 0
-	for _, aliasName := range expectedAliases {
-		indices := []string{}
-		writeIndices := []string{}
-		for indexName, aliases := range verifiedAliasInfo {
-			state, ok := aliases[aliasName]
-			if !ok || !state.Present {
-				continue
-			}
-
-			indices = append(indices, indexName)
-			if state.IsWriteIndex {
-				writeIndices = append(writeIndices, indexName)
-			}
-		}
-
-		if len(indices) == 0 {
-			result.MissingAliases = append(result.MissingAliases, aliasName)
-			continue
-		}
-
-		result.FoundAliases++
-		if len(writeIndices) != 1 {
-			result.InvalidWriteAlias = append(result.InvalidWriteAlias, fmt.Sprintf("%s (write_indices=%d)", aliasName, len(writeIndices)))
-			continue
-		}
-
-		sorted := append([]string{}, indices...)
-		sort.Slice(sorted, func(i, j int) bool {
-			gi := getOpensearchGeneration(sorted[i])
-			gj := getOpensearchGeneration(sorted[j])
-			if gi == gj {
-				return sorted[i] > sorted[j]
-			}
-			return gi > gj
-		})
-
-		latest := sorted[0]
-		if writeIndices[0] != latest {
-			result.InvalidWriteAlias = append(result.InvalidWriteAlias, fmt.Sprintf("%s (write=%s latest=%s)", aliasName, writeIndices[0], latest))
-		}
-	}
-
-	if len(result.MissingAliases) > 0 || len(result.InvalidWriteAlias) > 0 || result.FoundAliases != result.ExpectedAliases {
-		result.Success = false
-		result.Reason = "Opensearch alias verification failed after repair"
-		log.Printf("[WARNING] %s. expected_aliases=%d found_aliases=%d missing=%d invalid_write=%d", result.Reason, result.ExpectedAliases, result.FoundAliases, len(result.MissingAliases), len(result.InvalidWriteAlias))
-	} else {
-		result.Success = true
-		result.Reason = "Opensearch alias and index state repaired without data reindexing"
-	}
-
-	if ismEnabled {
-		appendBase := map[string]bool{}
-		for _, baseIndex := range GetOpensearchRolloverIndexes() {
-			appendBase[strings.ToLower(baseIndex)] = true
-		}
-
-		for _, aliasName := range expectedAliases {
-			bi := strings.TrimPrefix(strings.ToLower(aliasName), strings.ToLower(GetESIndexPrefix("")))
-			if !appendBase[bi] {
-				continue
-			}
-			retention := getOpensearchRetentionDays(bi)
-			policyID := fmt.Sprintf("%s-%s", ismPolicyName, aliasName)
-			ismReady, ismErr := ensureOpensearchISMRolloverPolicy(ctx, opensearchUrl, aliasName, rolloverConfig, retention, ismPolicyName)
-			if ismErr != nil {
-				log.Printf("[WARNING] Failed ensuring ISM rollover policy '%s' in prefix fix: %s", policyID, ismErr)
-				continue
-			}
-			if !ismReady {
-				continue
-			}
-			for indexName, aliases := range verifiedAliasInfo {
-				state, ok := aliases[aliasName]
-				if !ok || !state.Present {
-					continue
-				}
-
-				if err := ensureOpensearchIndexRolloverAliasSetting(ctx, opensearchUrl, indexName, aliasName); err != nil {
-					log.Printf("[WARNING] Failed ensuring rollover alias on %s for alias %s: %s", indexName, aliasName, err)
-					continue
-				}
-
-				if err := ensureOpensearchIndexISMPolicy(ctx, opensearchUrl, indexName, policyID); err != nil {
-					log.Printf("[WARNING] Failed attaching ISM policy '%s' to %s: %s", policyID, indexName, err)
-				}
-			}
-		}
-	}
-
-	return result, nil
-}
-
-func handleAliasCollisionMigration(foundClient opensearchapi.Client, opensearchUrl, sourceIndex, targetIndex string) (string, error) {
-	targetExists, err := checkOpensearchIndexExists(foundClient, opensearchUrl, targetIndex)
-	if err != nil {
-		return "", err
-	}
-
-	if !targetExists {
-		if err := createOpensearchIndex(foundClient, opensearchUrl, targetIndex); err != nil {
-			return "", err
-		}
-	}
-
-	sourceCount, err := getOpensearchIndexCount(foundClient, opensearchUrl, sourceIndex)
-	if err != nil {
-		return "", err
-	}
-
-	targetCount, err := getOpensearchIndexCount(foundClient, opensearchUrl, targetIndex)
-	if err != nil {
-		return "", err
-	}
-
-	if sourceCount > 0 && targetCount < sourceCount {
-		taskID, err := startOpensearchReindexTask(foundClient, opensearchUrl, sourceIndex, targetIndex)
-		if err != nil {
-			return "", err
-		}
-
-		return taskID, nil
-	}
-
-	if sourceCount > targetCount {
-		return "", fmt.Errorf("target count %d is lower than source count %d", targetCount, sourceCount)
-	}
-
-	if err := deleteOpensearchIndex(foundClient, opensearchUrl, sourceIndex); err != nil {
-		return "", err
-	}
-
-	isWrite := true
-	actions := []OpensearchAliasAction{
-		{
-			Add: &OpensearchAliasActionTarget{Index: targetIndex, Alias: sourceIndex, IsWriteIndex: &isWrite},
-		},
-	}
-
-	if err := updateOpensearchAliases(foundClient, opensearchUrl, actions); err != nil {
-		return "", err
-	}
-
-	return "", nil
-}
-
-func checkOpensearchIndexExists(foundClient opensearchapi.Client, opensearchUrl, indexName string) (bool, error) {
-	req, err := http.NewRequest("GET", fmt.Sprintf("%s/%s", opensearchUrl, indexName), nil)
-	if err != nil {
-		return false, err
-	}
-
-	resp, err := foundClient.Client.Transport.Perform(req)
-	if err != nil {
-		return false, err
-	}
-
-	body, readErr := io.ReadAll(resp.Body)
-	resp.Body.Close()
-	if readErr != nil {
-		return false, readErr
-	}
-
-	if resp.StatusCode == 404 {
-		return false, nil
-	}
-
-	if resp.StatusCode >= 300 {
-		return false, fmt.Errorf("failed checking index %s: %s", indexName, string(body))
-	}
-
-	return true, nil
-}
-
-func getOpensearchIndexCount(foundClient opensearchapi.Client, opensearchUrl, indexName string) (int64, error) {
-	req, err := http.NewRequest("GET", fmt.Sprintf("%s/%s/_count", opensearchUrl, indexName), nil)
-	if err != nil {
-		return 0, err
-	}
-
-	resp, err := foundClient.Client.Transport.Perform(req)
-	if err != nil {
-		return 0, err
-	}
-
-	body, readErr := io.ReadAll(resp.Body)
-	resp.Body.Close()
-	if readErr != nil {
-		return 0, readErr
-	}
-
-	if resp.StatusCode >= 300 {
-		return 0, fmt.Errorf("failed counting index %s: %s", indexName, string(body))
-	}
-
-	parsed := struct {
-		Count int64 `json:"count"`
-	}{}
-
-	if err := json.Unmarshal(body, &parsed); err != nil {
-		return 0, err
-	}
-
-	return parsed.Count, nil
-}
-
-func startOpensearchReindexTask(foundClient opensearchapi.Client, opensearchUrl, sourceIndex, targetIndex string) (string, error) {
-	payload := map[string]interface{}{
-		"source": map[string]interface{}{
-			"index": sourceIndex,
-		},
-		"dest": map[string]interface{}{
-			"index": targetIndex,
-		},
-		"conflicts": "proceed",
-	}
-
-	body, err := json.Marshal(payload)
-	if err != nil {
-		return "", err
-	}
-
-	req, err := http.NewRequest("POST", fmt.Sprintf("%s/_reindex?wait_for_completion=false", opensearchUrl), bytes.NewBuffer(body))
-	if err != nil {
-		return "", err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := foundClient.Client.Transport.Perform(req)
-	if err != nil {
-		return "", err
-	}
-
-	respBody, readErr := io.ReadAll(resp.Body)
-	resp.Body.Close()
-	if readErr != nil {
-		return "", readErr
-	}
-
-	if resp.StatusCode >= 300 {
-		lowerBody := strings.ToLower(string(respBody))
-		if strings.Contains(lowerBody, "resource_already_exists_exception") {
-			return "", nil
-		}
-
-		return "", fmt.Errorf("failed starting reindex %s -> %s: %s", sourceIndex, targetIndex, string(respBody))
-	}
-
-	parsed := struct {
-		Task string `json:"task"`
-	}{}
-
-	if err := json.Unmarshal(respBody, &parsed); err != nil {
-		return "", err
-	}
-
-	if strings.TrimSpace(parsed.Task) == "" {
-		return "", fmt.Errorf("reindex task missing in response")
-	}
-
-	return parsed.Task, nil
-}
-
-func deleteOpensearchIndex(foundClient opensearchapi.Client, opensearchUrl, indexName string) error {
-	req, err := http.NewRequest("DELETE", fmt.Sprintf("%s/%s", opensearchUrl, indexName), nil)
-	if err != nil {
-		return err
-	}
-
-	resp, err := foundClient.Client.Transport.Perform(req)
-	if err != nil {
-		return err
-	}
-
-	body, readErr := io.ReadAll(resp.Body)
-	resp.Body.Close()
-	if readErr != nil {
-		return readErr
-	}
-
-	if resp.StatusCode == 404 {
-		return nil
-	}
-
-	if resp.StatusCode >= 300 {
-		return fmt.Errorf("failed deleting index %s: %s", indexName, string(body))
-	}
-
-	return nil
-}
-
-type opensearchAliasState struct {
-	Present      bool
-	IsWriteIndex bool
-}
-
-// shuffleOwnedOpensearchIndexPatterns returns a comma-separated index pattern list
-// matching only the indexes Shuffle itself manages (GetOpensearchBaseIndexes,
-// combined with whatever SHUFFLE_OPENSEARCH_INDEX_PREFIX is configured - which
-// defaults to empty, i.e. no prefix). Scoping the cluster-wide _cat/indices and
-// _alias lookups below to this pattern lets a least-privilege OpenSearch role
-// grant indices:monitor/settings/get only on Shuffle's own indexes instead of
-// requiring visibility into every index in the cluster (which shared/enterprise
-// clusters with other tenants' indexes typically won't grant).
-func shuffleOwnedOpensearchIndexPatterns() string {
-	baseIndexes := GetOpensearchBaseIndexes()
-	patterns := make([]string, 0, len(baseIndexes))
-	for _, baseIndex := range baseIndexes {
-		patterns = append(patterns, strings.ToLower(GetESIndexPrefix(baseIndex))+"*")
-	}
-
-	return strings.Join(patterns, ",")
-}
-
-func getOpensearchAliases(foundClient opensearchapi.Client, opensearchUrl string) (map[string]map[string]opensearchAliasState, error) {
-	aliasReq, err := http.NewRequest("GET", fmt.Sprintf("%s/%s/_alias", opensearchUrl, shuffleOwnedOpensearchIndexPatterns()), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	aliasResp, err := foundClient.Client.Transport.Perform(aliasReq)
-	if err != nil {
-		return nil, err
-	}
-
-	aliasBody, err := io.ReadAll(aliasResp.Body)
-	if err != nil {
-		aliasResp.Body.Close()
-		return nil, err
-	}
-	aliasResp.Body.Close()
-
-	if aliasResp.StatusCode >= 300 {
-		return nil, fmt.Errorf("failed reading opensearch aliases: %s", string(aliasBody))
-	}
-
-	rawAliasInfo := OpensearchAliasResponse{}
-	if err := json.Unmarshal(aliasBody, &rawAliasInfo); err != nil {
-		return nil, err
-	}
-
-	aliasInfo := map[string]map[string]opensearchAliasState{}
-	type aliasDetails struct {
-		IsWriteIndex bool `json:"is_write_index,omitempty"`
-	}
-
-	for indexName, aliasEntry := range rawAliasInfo {
-		aliasInfo[indexName] = map[string]opensearchAliasState{}
-		for aliasName, aliasRaw := range aliasEntry.Aliases {
-			details := aliasDetails{}
-			_ = json.Unmarshal(aliasRaw, &details)
-			aliasInfo[indexName][aliasName] = opensearchAliasState{Present: true, IsWriteIndex: details.IsWriteIndex}
-		}
-	}
-
-	return aliasInfo, nil
-}
-
-// getOpensearchIndices resolves the concrete backing indices matching
-// Shuffle's own index patterns. Deliberately implemented via /_settings
-// rather than /_cat/indices: _cat/* endpoints are cluster-level actions in
-// OpenSearch's security plugin (requiring cluster:monitor/* even when the
-// path includes an index pattern) and, once granted, let the credential
-// query /_cluster/state or unscoped /_cat/indices directly to see every
-// index's name/mappings/settings cluster-wide - a real cross-tenant
-// metadata leak on a shared/multi-tenant cluster. /_settings (and /_alias
-// below), being genuine per-index API endpoints, are enforced per matched
-// index by the security plugin: a request for a pattern outside the role's
-// granted index_patterns is rejected outright.
-func getOpensearchIndices(foundClient opensearchapi.Client, opensearchUrl string) ([]string, error) {
-	indicesReq, err := http.NewRequest("GET", fmt.Sprintf("%s/%s/_settings?filter_path=*.settings.index.provided_name", opensearchUrl, shuffleOwnedOpensearchIndexPatterns()), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	indicesResp, err := foundClient.Client.Transport.Perform(indicesReq)
-	if err != nil {
-		return nil, err
-	}
-
-	indicesBody, err := io.ReadAll(indicesResp.Body)
-	if err != nil {
-		indicesResp.Body.Close()
-		return nil, err
-	}
-	indicesResp.Body.Close()
-
-	if indicesResp.StatusCode >= 300 {
-		return nil, fmt.Errorf("failed reading opensearch indices: %s", string(indicesBody))
-	}
-
-	parsedIndices := map[string]interface{}{}
-	if len(bytes.TrimSpace(indicesBody)) > 0 {
-		if err := json.Unmarshal(indicesBody, &parsedIndices); err != nil {
-			return nil, err
-		}
-	}
-
-	indices := []string{}
-	for indexName := range parsedIndices {
-		if strings.TrimSpace(indexName) != "" {
-			indices = append(indices, indexName)
-		}
-	}
-
-	return indices, nil
-}
-
-func selectOpensearchAliasTargets(expectedAlias, doubleAlias string, aliasInfo map[string]map[string]opensearchAliasState, allIndices []string) ([]string, string) {
-	candidateMap := map[string]bool{}
-
-	for indexName, aliases := range aliasInfo {
-		if aliases[expectedAlias].Present {
-			candidateMap[indexName] = true
-		}
-		if doubleAlias != "" && aliases[doubleAlias].Present {
-			candidateMap[indexName] = true
-		}
-	}
-
-	for _, indexName := range allIndices {
-		if indexName == expectedAlias || strings.HasPrefix(indexName, expectedAlias+"-") {
-			candidateMap[indexName] = true
-			continue
-		}
-
-		if doubleAlias != "" && (indexName == doubleAlias || strings.HasPrefix(indexName, doubleAlias+"-")) {
-			candidateMap[indexName] = true
-		}
-	}
-
-	targetIndices := []string{}
-	for indexName := range candidateMap {
-		targetIndices = append(targetIndices, indexName)
-	}
-
-	if len(targetIndices) == 0 {
-		return targetIndices, ""
-	}
-
-	sort.Slice(targetIndices, func(i, j int) bool {
-		gi := getOpensearchGeneration(targetIndices[i])
-		gj := getOpensearchGeneration(targetIndices[j])
-		if gi == gj {
-			return targetIndices[i] > targetIndices[j]
-		}
-		return gi > gj
-	})
-
-	writeIndex := ""
-	for _, indexName := range targetIndices {
-		if indexName == expectedAlias || strings.HasPrefix(indexName, expectedAlias+"-") {
-			writeIndex = indexName
-			break
-		}
-	}
-
-	if writeIndex == "" {
-		writeIndex = targetIndices[0]
-	}
-
-	return targetIndices, writeIndex
-}
-
-func getOpensearchGeneration(indexName string) int {
-	parts := strings.Split(indexName, "-")
-	if len(parts) < 2 {
-		return 0
-	}
-
-	generation := parts[len(parts)-1]
-	value, err := strconv.Atoi(generation)
-	if err != nil {
-		return 0
-	}
-
-	return value
-}
-
-func createOpensearchIndex(foundClient opensearchapi.Client, opensearchUrl, indexName string) error {
-	indexConfig := OpensearchIndexConfig{}
-	customConfig := strings.TrimSpace(os.Getenv("OPENSEARCH_INDEX_CONFIG"))
-	if customConfig != "" {
-		if err := json.Unmarshal([]byte(customConfig), &indexConfig); err != nil {
-			return fmt.Errorf("invalid OPENSEARCH_INDEX_CONFIG: %w", err)
-		}
-
-		if len(indexConfig.Aliases) > 0 {
-			indexConfig.Aliases = nil
-		}
-	}
-
-	if len(indexConfig.Settings) == 0 && len(indexConfig.Mappings) == 0 {
-		indexConfig = OpensearchIndexConfig{
-			Settings: map[string]interface{}{
-				"number_of_shards":   3,
-				"number_of_replicas": 1,
-				"refresh_interval":   "30s",
-			},
-			Mappings: map[string]interface{}{
-				"dynamic_templates": []map[string]interface{}{
-					{
-						"strings_as_keywords": map[string]interface{}{
-							"match_mapping_type": "string",
-							"mapping": map[string]interface{}{
-								"type": "keyword",
-							},
-						},
-					},
-				},
-			},
-		}
-	}
-
-	indexConfigJson, err := json.Marshal(indexConfig)
-	if err != nil {
-		return err
-	}
-
-	createReq, err := http.NewRequest("PUT", fmt.Sprintf("%s/%s", opensearchUrl, indexName), bytes.NewBuffer(indexConfigJson))
-	if err != nil {
-		return err
-	}
-	createReq.Header.Set("Content-Type", "application/json")
-
-	createResp, err := foundClient.Client.Transport.Perform(createReq)
-	if err != nil {
-		return err
-	}
-
-	createRespBody, err := io.ReadAll(createResp.Body)
-	if err != nil {
-		createResp.Body.Close()
-		return err
-	}
-	createResp.Body.Close()
-
-	if createResp.StatusCode >= 300 {
-		return fmt.Errorf("failed creating index %s: %s", indexName, string(createRespBody))
-	}
-
-	return nil
-}
-
-func updateOpensearchAliases(foundClient opensearchapi.Client, opensearchUrl string, actions []OpensearchAliasAction) error {
-	aliasActions := OpensearchAliasActionsRequest{Actions: actions}
-	aliasBody, err := json.Marshal(aliasActions)
-	if err != nil {
-		return err
-	}
-
-	aliasReq, err := http.NewRequest("POST", fmt.Sprintf("%s/_aliases", opensearchUrl), bytes.NewBuffer(aliasBody))
-	if err != nil {
-		return err
-	}
-	aliasReq.Header.Set("Content-Type", "application/json")
-
-	aliasResp, err := foundClient.Client.Transport.Perform(aliasReq)
-	if err != nil {
-		return err
-	}
-
-	aliasRespBody, err := io.ReadAll(aliasResp.Body)
-	if err != nil {
-		aliasResp.Body.Close()
-		return err
-	}
-	aliasResp.Body.Close()
-
-	if aliasResp.StatusCode >= 300 {
-		return fmt.Errorf("failed updating aliases: %s", string(aliasRespBody))
-	}
-
-	return nil
-}
-
-func HandleFixOpensearchPrefix(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in opensearch prefix fix: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Api authentication failed"}`))
-		return
-	}
-
-	if user.Role != "admin" {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Only admins or support can run this"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	result, err := FixOpensearchIndexPrefix(ctx)
-	if err != nil {
-		log.Printf("[ERROR] Failed fixing opensearch index prefix: %s", err)
-		result.Success = false
-		result.Reason = err.Error()
-		responseData, _ := json.Marshal(result)
-		resp.WriteHeader(500)
-		resp.Write(responseData)
-		return
-	}
-
-	responseData, err := json.Marshal(result)
-	if err != nil {
-		resp.WriteHeader(500)
-		resp.Write([]byte(`{"success": false, "reason": "Failed JSON parsing"}`))
-		return
-	}
-
-	resp.Header().Set("Content-Type", "application/json")
-	resp.WriteHeader(200)
-	resp.Write(responseData)
 }
 
 func RunOpensearchOps(ctx context.Context) (*opensearchapi.ClusterHealthResp, error) {
